pyproject.toml
[project]
name = "mothertoken"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=2.14.0",
    "huggingface_hub>=0.20.0",
    "tiktoken>=0.7.0",
    "transformers>=4.40.0",
    "pyyaml>=6.0.1",
    "anthropic>=0.25.0",
    "google-generativeai>=0.5.0",
]

[dependency-groups]
tests = [
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
